refactor: 결과 페이지 Suspense, use 도입 및 불필요한 상태와 훅 제거
